libnotify-0.7 drops support for notify_notification_new_with_status_icon() and only has notfy_notification_new()
Since it was checked for with g_module_symbol() that caused libnotify not to work and the tray plugin to fall back to the inferior experience of notify-send. But actually notify_notification_new_with_status_icon() wasn't used, it just was pointless checked for.

Patch also adds a comment explaining that notify_notification_new() lost its 4th argument, but passing an extra NULL is fine.
Submitted:

Owen Taylor ( otaylor ).